Move Atom class into parent Rule.h file
