+++ src/profile/Rule.h
@@ -14,10 +14,31 @@
 #include <tuple>
 #include <utility>
 
-#include "Atom.h"
-
 namespace souffle {
 namespace profile {
+
+/*
+ * Class to hold information about souffle Atom profile information
+ */
+class Atom {
+public:
+    const std::string identifier;
+    const std::string rule;
+    const size_t level;
+    const size_t frequency;
+
+    Atom(std::string identifier, std::string rule, size_t level, size_t frequency)
+            : identifier(std::move(identifier)), rule(std::move(rule)), level(level), frequency(frequency) {}
+
+    bool operator<(const Atom& other) const {
+        if (rule != other.rule) {
+            return rule < other.rule;
+        } else if (identifier != other.identifier) {
+            return identifier < other.identifier;
+        }
+        return level < other.level;
+    }
+};
 
 /*
  * Class to hold information about souffle Rule profile information
